+++ tests/Fixtures/Discriminator.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Liip\Serializer\Fixtures;
+
+use JMS\Serializer\Annotation as JMS;
+
+/**
+ * @JMS\Discriminator(field = "type", map = {
+ *     "first": "Tests\Liip\Serializer\Fixtures\DiscriminatorFirstChild",
+ *     "second": "Tests\Liip\Serializer\Fixtures\DiscriminatorSecondChild"
+ * })
+ */
+abstract class Discriminator
+{
+}